% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_functions.R
\name{filter_lowly_expressed_genes}
\alias{filter_lowly_expressed_genes}
\title{Filter out lowly expressed genes}
\usage{
filter_lowly_expressed_genes(
  data,
  cpm_saf = 1,
  sampels_per = 0.9,
  return_cpm = F
)
}
\arguments{
\item{data}{and ExpressionSet array or a count matrix}

\item{cpm_saf}{the threshold for cmp expression, (default 1)}

\item{sampels_per}{the threshold for percentage of samples expressing the gene, (default 0.9)}

\item{return_cpm}{a parameter which determine if to return the cpm or raw data (default F)}
}
\value{
a filtered cpm matrix or the filtered original data
}
\description{
The function accepts an ExpressionSet array or a count matrix and keeps
genes that express more then cpm_saf (default 1) in at least sampels_per (default 0.9).
The function default is to return a filtered cpm_matrix, if return_cpm is F the function reruns the  original data filtered
}
