% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_functions.R
\name{get_cpm}
\alias{get_cpm}
\title{Calculate cpm}
\usage{
get_cpm(data)
}
\arguments{
\item{data}{}
}
\value{
a cmp matrix
}
\description{
The function accepts an ExpressionSet array or a count matrix
and returns a counts per million matrix
}
