% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_functions.R
\name{read_expression_data}
\alias{read_expression_data}
\title{Read gene expression data}
\usage{
read_expression_data(
  folder_path = NULL,
  count_matrix_path = NULL,
  gene_annotation_path = NULL,
  sample_annotation_path = NULL
)
}
\arguments{
\item{folder_path}{a path to the folder that contains the files}

\item{count_matrix_path}{a path to the count matrix}

\item{gene_annotation_path}{a path to the gene annotation}

\item{sample_annotation_path}{a path to the sample annotation}
}
\value{
an ExpressionSet
}
\description{
This function reads the a count matrix a gene annotation file
and a sample annotation file and returns an expression set.
The function accepts a path to the folder that contains all the files, or a
separate path to each file.
If folder_path is provided the function identifies the files by their names:
counts file should contain count in the name,
gene_annotation_path should contain gene and anno separated by any chr
sample_annotation_path should contain sample and anno separated by any chr
}
