% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_functions.R
\name{run_differncial_expression}
\alias{run_differncial_expression}
\title{Run differential expression analysis using edgeR}
\usage{
run_differncial_expression(
  ExpressionSet_data,
  group_col_name,
  cpm_saf = 1,
  sampels_per = 0.9,
  use_tmm = T,
  create_plots = T,
  design = NULL
)
}
\arguments{
\item{ExpressionSet_data}{an ExpressionSet object}

\item{group_col_name}{a string with the group column name}

\item{cpm_saf}{the threshold for cmp expression, (default 1)}

\item{sampels_per}{the threshold for percentage of samples expressing the gene, (default 0.9)}

\item{use_tmm}{should tmm normalization be used, default is T}

\item{create_plots}{should plots be created, default is T}

\item{desian}{an optional design matrix for the model, if none is provided the model is based only on the group}
}
\value{
A data frame with gene expression results or a list with the following items
dif_exp_results: A data frame with gene expression results
norm_factor_diagnostic_plot: A plot of the normalization factor by group, used to determine if to use tmm normalization
top_10_exp_plot: A plot of the proportion of expression from the 10 ten most expressed genes, used to determine if to use tmm normalization
mds_plot: An mds plot of the results
volcano_plot: A volcano plot of the results
}
\description{
The function accepts an ExpressionSet_data and the name of the group column
Lowly expressed genes are filtered out (genes that express less then cpm_saf (default 1) in at least sampels_per (default 0.9) of the smallest group)
The function returns differential expression results and if create_plots is True the function also returns diagnostic plots, and go terms enrichmnet
}
