% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_functions.R
\name{test_if_file_can_be_read}
\alias{test_if_file_can_be_read}
\title{Identify file path for count, and annotations}
\usage{
test_if_file_can_be_read(
  files_in_folder_path,
  file_pattern,
  file_name,
  folder_path
)
}
\arguments{
\item{files_in_folder_path}{a list of file names}

\item{file_pattern}{the pattern of the file to search}

\item{file_name}{the file general name}

\item{folder_path}{the path}
}
\value{
file_path
}
\description{
The function identifies the path for the counts or annotation file in
the list of files in the provided folder path, if a matching file is not preset
or multiple matching files are found the function raises an error
}
