% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_functions.R
\name{top10prop}
\alias{top10prop}
\title{Get percentage of expression from the top 10 genes}
\usage{
top10prop(expression_vec)
}
\arguments{
\item{expression_vec}{}
}
\value{
a fraction
}
\description{
This function is a helper function of run_differential_expression
The function accepts a cpm_matrix colum and returns the proportion of expression
from the top 10 most expressed genes
}
